requirements.txt
streamlit
google-generativeai
python-docx
PyPDF2
python-dotenv
